requirements.txt
flask
python-dotenv
pandas
PyMuPDF
sentence-transformers
scikit-learn
rank-bm25
nltk
python-docx
google-generativeai
mistralai
dotenv